% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ittyBitty.R
\name{ittyBitty}
\alias{ittyBitty}
\title{make a new ittyBitty number}
\usage{
ittyBitty(k = double(), w = double())
}
\arguments{
\item{k}{numeric, the exponent}

\item{w}{numeric, the mantista

An \code{ittyBitty} is a representation of a numer as w * 2^k

A simply way to cast a number (e.g., \code{myNum})as an \code{ittyBitty} is to use \code{ittyBitty(k=1,w=myNum)}, 
The expected method of creating an \code{ittyBitty} is \code{exp_to_ittyBitty} which finds
\code{exp} of \code{x} as an \code{ittyBitty}. These can then be added together for numerical quadrature
and the results can be returned as a numeric with \code{log_as_numeric}. Doing this, far smaller
values can be represented than with doubles.}
}
\description{
make a new ittyBitty number
}
