% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ittyBitty.R
\name{[,ittyBitty,ANY,ANY,ANY-method}
\alias{[,ittyBitty,ANY,ANY,ANY-method}
\title{Setting subset and length methods}
\usage{
\S4method{[}{ittyBitty,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Setting subset and length methods
}
